Moved therioform model to a separate file
