+++ fabula_charsheet/assets/locals/en/classes.yaml
@@ -5,7 +5,7 @@ class_rogue: "Rogue"
 class_mutant: "Mutant"
 class_guardian: "Guardian"
 class_arcanist: "Arcanist"
-class_dark_blade: "Darkblade"
+class_dark_blade: "Dark Blade"
 class_elementalist: "Elementalist"
 class_entropist: "Entropist"
 class_fury: "Fury"
